# docs/swagger.yaml
basePath: /
definitions:
...
  input.AdminChangePasswordReq:
    properties:
      newPassword:
        description: 新密码
        example: "654321"
        maxLength: 32
        minLength: 6
        type: string
      oldPassword:
        description: 旧密码
        example: "123456"
        maxLength: 32
        minLength: 6
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  input.AdminDeleteReq:
    properties:
      id:
        description: 管理员ID
        example: 2
        type: integer
    required:
    - id
    type: object
  input.AdminListPageReq:
    properties:
      enable:
        description: 是否启用
        enum:
        - 0
        - 1
        example: 1
        type: integer
      gender:
        description: 性别
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      pageNo:
        description: 页码
        example: 1
        type: integer
      pageSize:
        description: 每页展示条数
        example: 20
        type: integer
      sortField:
        description: 排序字段
        example: points
        type: string
      sortOrder:
        description: 排序方向 ascend/descend
        enum:
        - ascend
        - descend
        example: descend
        type: string
      username:
        description: 用户名，支持模糊搜索
        example: admin
        type: string
    required:
    - pageNo
    - pageSize
    type: object
  input.AdminLoginReq:
    properties:
      captcha:
        description: 验证码
        type: string
      password:
        description: 密码
        example: "123456"
        maxLength: 32
        minLength: 6
        type: string
      username:
        description: 账号
        example: admin
        minLength: 1
        type: string
    required:
    - password
    - username
    type: object
  input.AdminRefreshReq:
    properties:
      token:
        description: refresh token
        example: Bearer xxxxxxxxxx
        type: string
    required:
    - token
    type: object
  input.AdminResetAdminPasswordReq:
    properties:
      id:
        description: 管理员ID
        example: 1
        type: integer
      password:
        description: 密码
        example: "123456"
        maxLength: 32
        minLength: 6
        type: string
    required:
    - id
    - password
    type: object
  input.AdminSaveReq:
    properties:
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 管理员ID
        example: 1
        type: integer
      password:
        description: 密码
        example: "123456"
        type: string
      roleIds:
        description: 角色ID组
        example:
        - 1
        - 2
        items:
          type: integer
        type: array
      username:
        description: 账号
        example: admin
        minLength: 1
        type: string
    required:
    - username
    type: object
  input.AdminSwitchRoleReq:
    properties:
      code:
        description: 角色标识
        example: SUPER_ADMIN
        type: string
    required:
    - code
    type: object
  input.AdminUpdateProfileReq:
    properties:
      address:
        description: 居住地址
        example: xxxxxxx
        type: string
      avatar:
        description: 头像地址
        example: https://cdn.hejunjie.life/avatars/oneadmin.jpeg
        type: string
      email:
        description: email
        example: xxxxx@xxx.com
        type: string
      gender:
        description: 性别
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      id:
        description: 管理员ID
        example: 1
        type: integer
      nickName:
        description: 名称
        example: test name
        type: string
    required:
    - id
    type: object
...
  resp.AdminCaptchaResp:
    properties:
      enabled:
        description: 是否启用验证码
        example: true
        type: boolean
    type: object
  resp.AdminDetailsProfileItem:
    properties:
      address:
        description: 居住地址
        example: xxxxxxx
        type: string
      avatar:
        description: 头像
        example: https://cdn.hejunjie.life/avatars/oneadmin.jpeg
        type: string
      email:
        description: email
        example: xxxxx@xxx.com
        type: string
      gender:
        description: 性别
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      id:
        description: 管理员ID
        example: 1
        type: integer
      nickName:
        description: 名称
        example: test name
        type: string
    type: object
  resp.AdminDetailsResp:
    properties:
      createdAt:
        description: 创建时间
        example: "2026-03-25 12:00:00"
        type: string
      currentRole:
        allOf:
        - $ref: '#/definitions/resp.AdminDetailsRoleItem'
        description: 当前角色信息
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 管理员ID
        example: 1
        type: integer
      profile:
        allOf:
        - $ref: '#/definitions/resp.AdminDetailsProfileItem'
        description: 个人资料
      roles:
        description: 拥有角色列表
        items:
          $ref: '#/definitions/resp.AdminDetailsRoleItem'
        type: array
      updatedAt:
        description: 更新时间
        example: "2026-03-25 12:00:00"
        type: string
      username:
        description: 账号
        example: admin
        type: string
    type: object
  resp.AdminDetailsRoleItem:
    properties:
      code:
        description: 角色标识
        example: SUPER_ADMIN
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      id:
        description: 角色ID
        example: 1
        type: integer
      name:
        description: 角色名称
        example: 超级管理员
        type: string
    type: object
  resp.AdminListPageItem:
    properties:
      address:
        description: 居住地址
        example: xxxxxxx
        type: string
      avatar:
        description: 头像地址
        example: https://cdn.hejunjie.life/avatars/oneadmin.jpeg
        type: string
      createdAt:
        description: 创建时间
        type: string
      email:
        description: email
        example: xxxxx@xxx.com
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      gender:
        description: 性别
        enum:
        - 0
        - 1
        - 2
        example: 1
        type: integer
      id:
        description: 管理员ID
        example: 1
        type: integer
      roles:
        description: 拥有角色列表
        items:
          $ref: '#/definitions/resp.AdminDetailsRoleItem'
        type: array
      updatedAt:
        description: 更新时间
        type: string
      username:
        description: 账号
        example: admin
        type: string
    type: object
  resp.AdminListPageResp:
    properties:
      pageData:
        description: 当前页码数据
        items:
          $ref: '#/definitions/resp.AdminListPageItem'
        type: array
      total:
        description: 总计条数
        example: 100
        type: integer
    type: object
  resp.AdminLoginResp:
    properties:
      accessToken:
        description: access token
        example: Bearer xxxxxxxxxx
        type: string
      refreshToken:
        description: refresh token
        example: Bearer xxxxxxxxxx
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: 何俊杰
  description: BiliLive Assistant 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BiliLive Assistant API
  version: "1.0"
paths:
  /api/admin/admin/delete:
    post:
      description: 根据管理员 ID 删除管理员账号（通常不允许删除当前登录管理员）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 删除参数（管理员 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminDeleteReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除管理员
      tags:
      - 管理员
  /api/admin/admin/list:
    post:
      description: 分页获取管理员列表，支持按用户名、性别、启用状态进行筛选
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminListPageReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminListPageResp'
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询管理员列表
      tags:
      - 管理员
  /api/admin/admin/save:
    post:
      description: 根据是否传入 ID 创建或更新管理员账号：未传 ID 时为创建，传入 ID 时为更新，同时可设置账号状态及绑定角色
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 管理员保存参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminSaveReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新管理员
      tags:
      - 管理员
  /api/admin/admin/update-password:
    post:
      description: 管理员根据管理员 ID 重置指定账号的登录密码（无需旧密码，通常需要具备管理员管理权限）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 重置密码参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminResetAdminPasswordReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 重置管理员密码
      tags:
      - 管理员
  /api/admin/admin/update-profile:
    post:
      description: 管理员更新个人资料信息（如昵称、头像、性别、邮箱等），支持部分字段更新
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 管理员个人资料参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminUpdateProfileReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新管理员个人资料
      tags:
      - 管理员
...
  /api/admin/auth/captcha:
    post:
      description: 获取当前系统是否已配置并启用了 altcha 验证码，前端可根据此状态决定是否展示验证码组件
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminCaptchaResp'
              type: object
      summary: 获取验证码启用状态
      tags:
      - 认证
  /api/admin/auth/change-password:
    post:
      description: 管理员在已登录状态下通过旧密码验证后修改登录密码，修改成功后建议重新登录以获取新的凭证
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 修改密码参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminChangePasswordReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改登录密码
      tags:
      - 认证
  /api/admin/auth/detail:
    post:
      description: 获取当前登录管理员的详细信息，包括基础信息、个人资料、拥有角色列表及当前角色信息
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminDetailsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取当前管理员信息
      tags:
      - 认证
  /api/admin/auth/login:
    post:
      description: 管理员通过账号和密码登录系统，登录成功后返回 access_token 和 refresh_token，用于后续接口鉴权
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 登录参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminLoginReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      summary: 管理员登录
      tags:
      - 认证
  /api/admin/auth/logout:
    post:
      description: 管理员退出当前登录状态，使当前 access_token 失效（通常用于前端主动登出）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 管理员退出登录
      tags:
      - 认证
  /api/admin/auth/refresh:
    post:
      description: 使用 refresh_token 刷新登录状态，获取新的 access_token 和 refresh_token，用于延长会话有效期
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 刷新凭证参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminRefreshReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      summary: 刷新登录凭证
      tags:
      - 认证
  /api/admin/auth/switch-role:
    post:
      description: 管理员在已登录状态下切换当前角色，根据目标角色重新生成 access_token 和 refresh_token，用于更新权限上下文
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 切换角色参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AdminSwitchRoleReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AdminLoginResp'
              type: object
      security:
      - BearerAuth: []
      summary: 切换管理员角色
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"